pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-server-servicenow"
version = "0.4.0"
description = "MCP server for ServiceNow - table, CMDB, system, and update set operations"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Josh Schuller"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=3.0.2,<4",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "authlib>=1.3.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://github.com/jschuller/mcp-server-servicenow"
Repository = "https://github.com/jschuller/mcp-server-servicenow"
Issues = "https://github.com/jschuller/mcp-server-servicenow/issues"

[project.scripts]
mcp-server-servicenow = "servicenow_mcp.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-mock>=3.0", "pytest-asyncio>=0.23.0", "ruff>=0.4.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/servicenow_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
